Moved Neo4jGraph and associated objects to a new file
